backend/requirements.txt
fastapi
uvicorn
gitpython
radon
bandit
pip-audit
openai
python-multipart